feat(schematics): adds schematics to install and scaffold initial files
this schematic provides setup of serenity with protractor/cucumber. it installs all required
dependencies, configures protractor and scaffolds the code files for actors and cucumber timeouts.
the schematic accepts parameters for selecting the testingFramework (currently only cucumber),
browserIntegration (currently only protractor) as well as the (angular) project name to which
serenity should be added (defaults to the project set as defaultProject)
and a directory to use for the e2e testing code (defaults to features).

 #359